Add datatable and sweetalert

+++ app/Http/Controllers/User/UserController.php
@@ -9,6 +9,9 @@ use Illuminate\Http\Request;
 use Illuminate\View\View;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Auth;
+use Symfony\Component\HttpFoundation\Response;
+use Yajra\DataTables\Facades\DataTables;
 
 use App\Models\User;
 
@@ -25,11 +28,23 @@ class UserController extends Controller
 
 		// users
 		// paginate with query
-		$users = User::where('name', 'LIKE', '%'.$searchVal.'%')->whereNot('id', auth()->user()->id)->paginate(5)->withQueryString();
+		$users = User::where('name', 'LIKE', '%' . $searchVal . '%')->whereNot('id', auth()->user()->id)->paginate(5)->withQueryString();
 
 		return view('user.index', compact('users', 'searchVal'));
 	}
 
+	public function showTable(Request $request)
+	{
+		if ($request->ajax()) {
+
+			$users = User::where('id', '!=', Auth::id())->select('id', 'name', 'username', 'role', 'contact_number');
+
+			return DataTables::of($users)
+				->addColumn('action', 'user.table-buttons')
+				->rawColumns(['action', 'created_at'])
+				->toJson();
+		}
+	}
 	/**
 	 * Show the form for creating a new resource.
 	 */
@@ -56,8 +71,8 @@ class UserController extends Controller
 		$user->password 			= Hash::make($validated['password']);
 		$user->save();
 
-		// redirect to users page
-		return redirect()->route('users.index')->with('status', 'User has been successfully added.');
+		toast('User has been successfully added.', 'success');
+		return redirect()->route('users.index');
 	}
 
 	/**
@@ -98,20 +113,23 @@ class UserController extends Controller
 
 		$user->update();
 
-		// redirect to users page
-		return redirect()->route('users.index')->with('status', 'User has been successfully updated.');
+		toast('User has been successfully updated.', 'success');
+		return redirect()->route('users.index');
 	}
 
 	/**
 	 * Remove the specified resource from storage.
 	 */
-	public function destroy(User $user): RedirectResponse
+	public function destroy(Request $request, User $user)
 	{
-		// delete User
-		$user->delete();
+		if ($request->ajax()) {
 
-		// redirect to users page
-		return redirect()->route('users.index')->with('status', 'User has been successfully deleted.');
+			$user->delete();
 
+			return response()->json([
+				'success'  => true,
+				'message'  => 'User has been successfully deleted.'
+			], Response::HTTP_OK);
+		}
 	}
 }
